Adopt and initialize Flask-Migrate
The application factory no longer calls db.create_all() because that
would interfere with migration autodetection. The consequence is that
the database tables will not be created unless you run the migrations.

This may lead to conflicts, since Flask-Admin relies on the existence
of the tables. The -A flag was added to prevent this issue; it
prevents the application factory from creating the admin instance.
When the database is not up-to-date yet, pass this flag to manage.py.